# openapi.yaml
openapi: 3.1.0
info:
  title: Axon
  version: 0.1.0
  summary: Universal API gateway for autonomous agents
  description: |
    Axon proxies calls to a catalog of paid APIs and settles payment per
    request in USDC on Base. One endpoint, one wallet, per-request pricing,
    automatic cache, auto-refund on upstream failures.

    Two authentication modes:
      1. **Prepaid wallet** — default. Deposit USDC, then authenticate with
         `x-api-key`. We debit from balance per call.
      2. **x402 native** — no API key. Responds 402 with payment
         requirements, client pays on-chain per call. Use `/x402/v1/call/*`.

    Every response carries `x-axon-cost-usdc`, `x-axon-cache`,
    `x-axon-latency-ms`, and `x-request-id`.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Axon
    url: https://axon.dev
servers:
  - url: https://api.axon.dev
    description: Production (hosted)
  - url: http://localhost:3000
    description: Local development
...
security:
  - apiKey: []
paths:
...
  /v1/usage:
    get:
      tags: [Usage]
      summary: Your total usage and cache hit rate
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: api
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_requests: { type: integer }
                  cache_hits: { type: integer }
                  cache_hit_rate: { type: number }
                  total_spent_usdc: { type: string }
...
  /v1/admin/policy/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Admin]
      summary: Read a user's policy
      security: [{ adminKey: [] }]
      responses:
        '200':
          description: Policy
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string }
                  policy:
                    oneOf:
                      - $ref: '#/components/schemas/Policy'
                      - { type: 'null' }
    put:
      tags: [Admin]
      summary: Upsert a user's policy
      security: [{ adminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Policy' }
      responses:
        '200': { description: OK }
    delete:
      tags: [Admin]
      summary: Remove a user's policy
      security: [{ adminKey: [] }]
      responses:
        '200': { description: OK }
  /v1/admin/settlements:
    get:
      tags: [Admin]
      summary: List settlements
      security: [{ adminKey: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, paid, reconciled] }
      responses:
        '200': { description: List }
  /v1/admin/settlements/run:
    post:
      tags: [Admin]
      summary: Trigger a settlement run
      security: [{ adminKey: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                start: { type: string, format: date-time }
                end: { type: string, format: date-time }
      responses:
        '200': { description: Results }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Starts with `ax_live_`
    adminKey:
      type: apiKey
      in: header
      name: x-admin-key
...
  schemas:
...
    Policy:
      type: object
      properties:
        allow_apis:
          type: array
          items: { type: string }
        deny_apis:
          type: array
          items: { type: string }
        daily_budget_micro: { type: string, description: bigint as string }
        monthly_budget_micro: { type: string }
        max_request_cost_micro: { type: string }
        per_api_daily_micro:
          type: object
          additionalProperties: { type: string }
        exclude_cache_from_budget: { type: boolean }
        label: { type: string }